Split vpc code into api/services/db layers
